refactor(offline): 整理离线填报模块文件结构
- 将离线填报相关 Activity 移入 container.offlineform 包，减少 container 根包下的业务页面堆积
- 同步 AndroidManifest 中离线填报 Activity 的声明路径
- 更新 MainActivity 和 OfflinePlusCardAdapter 对离线填报 Activity 的 import
- 统一离线填报 layout 资源命名，使用 offline_form_* 前缀聚合模块相关布局
- 更新 CustomFormActivity、OfflinePlusListActivity、OfflineProjectRecordActivity 和 OfflinePlusCardAdapter 的 R.layout 引用
- 新增 strings_offline_form.xml，集中存放 offline_* 文案
- 从主 strings.xml 移除 offline_* 文案，保留原资源名不变，避免影响调用方
