+++ app/src/main/AndroidManifest.xml
@@ -142,13 +142,13 @@
         <activity
             android:name=".proxy.support.biometric.BiometricAuthActivity"
             android:exported="false" />
-        <activity android:name=".CustomFormActivity"
+        <activity android:name=".offlineform.CustomFormActivity"
             android:exported="false" />
-        <activity android:name=".OfflinePlusListActivity"
+        <activity android:name=".offlineform.OfflinePlusListActivity"
             android:exported="false" />
-        <activity android:name=".OfflineProjectRecordActivity"
+        <activity android:name=".offlineform.OfflineProjectRecordActivity"
             android:exported="false" />
-        <activity android:name=".OfflineFormConfigDetailActivity"
+        <activity android:name=".offlineform.OfflineFormConfigDetailActivity"
             android:exported="false" />
 
         <provider
