+++ app/src/main/java/com/huozige/lab/container/proxy/support/offlinecustomform/OfflinePlusCardAdapter.java
@@ -13,7 +13,7 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.huozige.lab.container.OfflineProjectRecordActivity;
+import com.huozige.lab.container.offlineform.OfflineProjectRecordActivity;
 import com.huozige.lab.container.R;
 import com.huozige.lab.container.proxy.support.offlinecustomform.helper.OfflineComputedHelper;
 import com.huozige.lab.container.proxy.support.offlinecustomform.model.OfflineFormDefinitionIndexItem;
@@ -40,7 +40,7 @@ public class OfflinePlusCardAdapter extends RecyclerView.Adapter<OfflinePlusCard
     @Override
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View view = LayoutInflater.from(parent.getContext())
-                .inflate(R.layout.offline_card, parent, false);
+                .inflate(R.layout.offline_form_card, parent, false);
         return new ViewHolder(view);
     }
