+++ app/src/main/res/layout/offline_form_list_activity.xml
