+++ app/src/main/res/layout/offline_form_project_record_activity.xml
